add rule to compress mater agg
