+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/analysis/NormalizeAggregate.java
@@ -126,95 +126,7 @@ public class NormalizeAggregate implements RewriteRuleFactory, NormalizeToSlot {
                         .toRule(RuleType.NORMALIZE_AGGREGATE));
     }
 
-    private boolean canCompress(Expression expression) {
-        DataType type = expression.getDataType();
-        if (type instanceof CharacterType) {
-            CharacterType ct = (CharacterType) type;
-            if (ct.getLen() < 7) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /*
-    example:
-    [support] select sum(v) from t group by substring(k, 1,2)
-    [not support] select substring(k, 1,2), sum(v) from t group by substring(k, 1,2)
-    [support] select k, sum(v) from t group by k
-    [not support] select substring(k, 1,2), sum(v) from t group by k
-    [support]  select A as B from T group by A
-    */
-    private Set<Expression> getEncodableGroupByExpressions(LogicalAggregate<Plan> aggregate) {
-        Set<Expression> encodableGroupbyExpressions = Sets.newHashSet();
-        Set<Slot> slotShouldNotEncode = Sets.newHashSet();
-        for (NamedExpression ne : aggregate.getOutputExpressions()) {
-            if (ne instanceof Alias) {
-                Expression child = ((Alias) ne).child();
-                //support: select A as B from T group by A
-                if (!(child instanceof SlotReference)) {
-                    slotShouldNotEncode.addAll(child.getInputSlots());
-                }
-            }
-        }
-        for (Expression gb : aggregate.getGroupByExpressions()) {
-            if (canCompress(gb)) {
-                boolean encodable = true;
-                for (Slot gbs : gb.getInputSlots()) {
-                    if (slotShouldNotEncode.contains(gbs)) {
-                        encodable = false;
-                        break;
-                    }
-                }
-                if (encodable) {
-                    encodableGroupbyExpressions.add(gb);
-                }
-            }
-        }
-        return encodableGroupbyExpressions;
-    }
-
-    private LogicalAggregate<Plan> encode(LogicalAggregate<Plan> aggregate, Optional<LogicalHaving<?>> having) {
-        List<Alias> encodedExpressions = Lists.newArrayList();
-        Set<Expression> encodableGroupByExpressions = getEncodableGroupByExpressions(aggregate);
-        if (!encodableGroupByExpressions.isEmpty()) {
-            List<Expression> newGroupByExpressions = Lists.newArrayList();
-            List<Alias> encodedGroupByExpressions = Lists.newArrayList();
-            for (Expression gp : aggregate.getGroupByExpressions()) {
-                if (encodableGroupByExpressions.contains(gp)) {
-                    Alias alias = new Alias(new EncodeAsBigInt(gp));
-                    newGroupByExpressions.add(alias);
-                    encodedExpressions.add(alias);
-                } else {
-                    newGroupByExpressions.add(gp);
-                }
-            }
-            List<NamedExpression> newOutput = Lists.newArrayList();
-            for (NamedExpression ne : aggregate.getOutputExpressions()) {
-                if (ne instanceof SlotReference && encodableGroupByExpressions.contains(ne)) {
-                    newOutput.add(new Alias(ne.getExprId(), new AnyValue(ne), ne.getName()));
-                } else if (ne instanceof Alias && encodableGroupByExpressions.contains(((Alias) ne).child())) {
-                    Expression child = ((Alias) ne).child();
-                    Preconditions.checkArgument(child instanceof SlotReference,
-                            "encode %s failed, not a slot", child);
-                    newOutput.add(new Alias(((SlotReference) child).getExprId(), new AnyValue(child),
-                            "any_value(" + child + ")"));
-                } else {
-                    newOutput.add(ne);
-                }
-            }
-            newOutput.addAll(encodedExpressions);
-            aggregate = aggregate.withGroupByAndOutput(newGroupByExpressions, newOutput);
-        }
-        return aggregate;
-    }
-
     private LogicalPlan normalizeAgg(LogicalAggregate<Plan> aggregate, Optional<LogicalHaving<?>> having) {
-        aggregate = encode(aggregate, having);
-        return normalizeAggInner(aggregate, having);
-    }
-
-    private LogicalPlan normalizeAggInner(LogicalAggregate<Plan> aggregate, Optional<LogicalHaving<?>> having) {
         // The LogicalAggregate node may contain window agg functions and usual agg functions
         // we call window agg functions as window-agg and usual agg functions as trivial-agg for short
         // This rule simplify LogicalAggregate node by:
